requirements.txt
Flask==1.1.1
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.13.2
MarkupSafe==1.1.1
numpy==1.16.5
pandas==0.25.1
scikit-learn==0.23.2
scipy==1.4.1
virtualenv==20.0.25
Werkzeug==0.16.0
xgboost==1.4.2
cassandra-driver==3.25.0
